Move startActivity extension to common module Use it where we can
